server/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.120.4
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
google-auth==2.42.1
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.3
rsa==4.9.1
sentry-sdk==2.43.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.3
tenacity==9.1.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
